backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Computer vision and ML
ultralytics==8.3.221
opencv-python==4.10.0.84
numpy==2.2.0
# Image processing
Pillow==11.0.0

# AI and TTS services
fish-audio-sdk==2025.6.3
groq==0.11.0
python-dotenv==1.0.1

